requirements.txt
ailment==9.2.47
angr==9.2.47
ansi2html==1.8.0
archinfo==9.2.47
asttokens==2.0.8
asyncio==3.4.3
backcall==0.2.0
bitstring==3.1.9
cachetools==5.2.0
capstone==4.0.2
cffi==1.15.1
claripy==9.2.47
cle==9.2.47
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
CppHeaderParser==2.7.4
decorator==5.1.1
dominate==2.7.0
dpkt==1.9.8
executing==1.0.0
future==0.18.2
gitdb==4.0.9
GitPython==3.1.27
ipython==8.5.0
itanium-demangler==1.1
jedi==0.18.1
markdown-it-py==2.2.0
matplotlib-inline==0.1.6
mdurl==0.1.2
mpmath==1.2.1
mulpyplexer==0.9
nampa==0.1.1
networkx==2.8.6
parso==0.8.3
pefile==2022.5.30
pexpect==4.8.0
pickleshare==0.7.5
plumbum==1.7.2
ply==3.11
progressbar2==4.0.0
prompt-toolkit==3.0.31
protobuf==4.21.6
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pyelftools==0.29
Pygments==2.13.0
PySMT==0.9.6.dev21
python-utils==3.3.3
pyvex==9.2.47
rich==13.3.4
rpyc==5.2.3
shellingham==1.5.0
six==1.16.0
smmap==5.0.0
sortedcontainers==2.4.0
stack-data==0.5.0
sympy==1.11.1
traitlets==5.4.0
typer==0.6.1
typing_extensions==4.3.0
unicorn==2.0.1.post1
wcwidth==0.2.5
z3-solver==4.10.2.0
